+++ src/app/masterfile/ancillarysection/ancillarysection.component.ts
@@ -1,5 +1,6 @@
 import { DatePipe } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
+import { ConfirmationService } from 'primeng/api';
 import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
 import { AncillarySection } from 'src/app/models/ancillarysection.model';
 import { AncillarysectionService } from 'src/app/services/ancillarysection.service';
@@ -9,22 +10,25 @@ import { PopupAncillarysectionComponent } from '../popup/popup-ancillarysection/
   selector: 'app-ancillarysection',
   templateUrl: './ancillarysection.component.html',
   styleUrls: ['./ancillarysection.component.css'],
-  providers: [DialogService]
+  providers: [DialogService, ConfirmationService]
 })
 export class AncillarysectionComponent implements OnInit {
   searchkey: ""
   ref: DynamicDialogRef;
-  ancillarysection : AncillarySection;
-  prevAncillarySectionList : AncillarySection[];
-  ancillarysectionList : AncillarySection[];
-  selectedAncillarySection : AncillarySection[];
+  ancillarysection: AncillarySection;
+  prevAncillarySectionList: AncillarySection[];
+  ancillarysectionList: AncillarySection[];
+  selectedAncillarySection: AncillarySection[];
+  responsemessage: string;
+  headermessage: string;
+  displayResponsive: boolean = false;
 
-  constructor(private ancillarysectionservice : AncillarysectionService, private dialogService : DialogService, private datePipe : DatePipe) { }
+  constructor(private ancillarysectionservice: AncillarysectionService, private confirmationService: ConfirmationService, private dialogService: DialogService, private datePipe: DatePipe) { }
 
   ngOnInit(): void {
     this.fetchData();
   }
-  fetchData(){
+  fetchData() {
     this.ancillarysectionservice.getAncillarySection().subscribe({
       next: (result: AncillarySection[]) => {
         this.ancillarysectionList = result;
@@ -95,22 +99,29 @@ export class AncillarysectionComponent implements OnInit {
     })
   }
   removeAncillarySectionRecord(ancillarysection: AncillarySection) {
-    this.ancillarysectionservice.deleteAncillarySection(ancillarysection.id).subscribe({
-      next: (result: boolean) => {
-        result;
-        this.ancillarysectionList.forEach(element => {
-          if (ancillarysection.id == element.id) {
-            element.status = false;
-          }
+    this.confirmationService.confirm({
+      message: `Are you sure you want to delete the record?`,
+      header: 'Confirm',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.ancillarysectionservice.deleteAncillarySection(ancillarysection.id).subscribe({
+          next: (result: boolean) => {
+            result;
+            this.ancillarysectionList.forEach(element => {
+              if (ancillarysection.id == element.id) {
+                element.status = false;
+              }
 
+            });
+          },
+          error: (err: any) => {
+            console.log(err);
+          },
+          complete: () => {
+            console.log('complete');
+            this.prevAncillarySectionList = this.ancillarysectionList.filter(x => x.status);
+          }
         });
-      },
-      error: (err: any) => {
-        console.log(err);
-      },
-      complete: () => {
-        console.log('complete');
-        this.prevAncillarySectionList = this.ancillarysectionList.filter(x => x.status);
       }
     });
   }
